+++ src/names.c
@@ -2721,6 +2721,20 @@ static type_t solve_array_ref(nametab_t *tab, tree_t ref)
    else
       elem_type = type_new(T_NONE);
 
+   if (standard() >= STD_08 && type_kind(base_type) == T_SUBTYPE) {
+      const int ncon = type_constraints(base_type);
+      if (ncon > 1) {
+         assert(type_is_unconstrained(elem_type));
+
+         type_t sub = type_new(T_SUBTYPE);
+         type_set_base(sub, elem_type);
+         for (int i = 1; i < ncon; i++)
+            type_add_constraint(sub, type_constraint(base_type, i));
+
+         elem_type = sub;
+      }
+   }
+
    tree_set_type(ref, elem_type);
    return elem_type;
 }
@@ -2741,6 +2755,12 @@ static type_t solve_array_slice(nametab_t *tab, tree_t slice)
    type_t slice_type = type_new(T_SUBTYPE);
    type_set_base(slice_type, base_type);
    type_add_constraint(slice_type, constraint);
+
+   if (standard() >= STD_08 && type_kind(base_type) == T_SUBTYPE) {
+      const int ncon = type_constraints(base_type);
+      for (int i = 1; i < ncon; i++)
+         type_add_constraint(slice_type, type_constraint(base_type, i));
+   }
 
    tree_set_type(slice, slice_type);
    return slice_type;